+++ packages/plugin-javascript-fixtures/src/index.ts
@@ -1,0 +1,4 @@
+import { plugin } from './plugin';
+
+export * from './plugin';
+export default plugin;