+++ packages/plugin-jest/templates/jest-plugin.js
@@ -1,4 +1,5 @@
 const { sextantScenarios } = require('./sextant-types.generated');
+const { mockSextantEvent } = require('./sextant-fixture-mock.generated');
 
 /**
  * Describe a feature of Sextant in a jest describe block
@@ -13,6 +14,8 @@ export const describeSextantFeature = (featureName, testFunc) => {
         testsRegistered.push(name);
         test(name, ...args);
       },
+      mockEvent: (eventName, event) =>
+        mockSextantEvent(featureName, eventName, event),
       testCoverage: () => {
         test('Coverage of all Sextant scenarios', () => {
           desiredScenarios.forEach((scenarioName) => {